Make iceberg-catalog build and pass tests; Make iceberg-handler depend on iceberg-catalog
