+++ packaging/pom.xml
@@ -230,6 +230,11 @@
     </dependency>
     <dependency>
       <groupId>org.apache.hive</groupId>
+      <artifactId>hive-iceberg-catalog</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.hive</groupId>
       <artifactId>hive-iceberg-handler</artifactId>
       <version>${project.version}</version>
     </dependency>
